Removing example JS code from JS validator facade
