+++ src/main/java/org/raml/amf/shapes/Validator.java
@@ -1,0 +1,7 @@
+package org.raml.amf.shapes;
+
+/**
+ * Created by antoniogarrote on 09/05/2017.
+ */
+public class Validator {
+}
